pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pm_predictor"
version = "0.1.0"
description = "PM2.5 forecasting to avoid runner's nose"
dependencies = [
  "pandas>=1.3",
  "scikit-learn>=1.0",
  "google-cloud-storage>=2.0",
  "joblib>=1.0"
]
readme = "README.md"
requires-python = ">=3.9"

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points."console_scripts"]
train-aq = "src.train:main"